refactor: menu component
Moved component files into separate folder
